+++ apps/notifications/services.py
@@ -3,6 +3,7 @@ from typing import List
 from django.apps import apps
 from django.template.loader import render_to_string
 from django.utils import translation
+from django.utils.safestring import mark_safe
 from guest_user.functions import is_guest_user
 
 from apps.users.emails import EmailAplus as Email
@@ -56,6 +57,7 @@ class NotificationEmail(Email):
                         event_name=context.get("event_name"),
                         project_name=context.get("project_name"),
                         project_type=context.get("project_type"),
+                        organisation_name=context.get("organisation_name", ""),
                         article=context.get("article", ""),
                         content_type=context.get("content_type", ""),
                         content_type_display=context.get("content_type_display", ""),
@@ -84,11 +86,22 @@ class NotificationEmail(Email):
                         receiver_name=receiver.username
                     )
                 if "reason" in context:
+                    organisation_path = context.get("organisation_path", "")
+                    organisation_url = (
+                        self.get_host() + organisation_path if organisation_path else ""
+                    )
+                    unfollow_path = context.get("unfollow_path", "")
+                    unfollow_url = (
+                        self.get_host() + unfollow_path if unfollow_path else ""
+                    )
                     context["reason"] = context["reason"].format(
                         receiver_email=receiver.email,
                         organisation_name=context.get("organisation_name", ""),
                         site_name=site_name,
+                        organisation_url=organisation_url,
+                        unfollow_url=unfollow_url,
                     )
+                    context["reason"] = mark_safe(context["reason"])
 
             return super().render(template_name, context)